+++ .idea/workspace.xml
@@ -7,10 +7,8 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="d6c24df8-0d7f-49f5-b43f-443890cc9009" name="Default Changelist" comment="">
-      <change beforePath="$PROJECT_DIR$/.idea/vcs.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/vcs.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/out/production/Laba_5/Main$1.class" beforeDir="false" afterPath="$PROJECT_DIR$/out/production/Laba_5/Main$1.class" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/out/production/Laba_5/Main.class" beforeDir="false" afterPath="$PROJECT_DIR$/out/production/Laba_5/Main.class" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/CommandReader.java" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/src/Main.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/Main.java" afterDir="false" />
     </list>
     <option name="EXCLUDED_CONVERTED_TO_IGNORED" value="true" />
@@ -87,6 +85,14 @@
       <workItem from="1559208991459" duration="613000" />
       <workItem from="1559210566437" duration="5246000" />
     </task>
+    <task id="LOCAL-00001" summary="1">
+      <created>1569182870479</created>
+      <option name="number" value="00001" />
+      <option name="presentableId" value="LOCAL-00001" />
+      <option name="project" value="LOCAL" />
+      <updated>1569182870479</updated>
+    </task>
+    <option name="localTasksCounter" value="2" />
     <servers />
   </component>
   <component name="TimeTrackingManager">
@@ -105,5 +111,9 @@
         </entry>
       </map>
     </option>
+  </component>
+  <component name="VcsManagerConfiguration">
+    <MESSAGE value="1" />
+    <option name="LAST_COMMIT_MESSAGE" value="1" />
   </component>
 </project>
