+++ src/Main.java
@@ -33,6 +33,7 @@ public class Main {
     private static AreaComparator objAreaComparator = new AreaComparator();
     private static String collectionPath;
     private static List<House> list = Collections.emptyList();
+    private static boolean isExit;
 
     private static House fromCSV(String line, String separator) {
         try {
@@ -112,7 +113,152 @@ public class Main {
 
         System.out.println("--- START ---");
         System.out.println("введите команду: ");
-        CommandReader.goIn();
+        try {
+            while (!isExit) {
+                System.out.print("> ");
+                String command = scanner.next();
+                switch (command.toLowerCase()) {
+                    case CMD_EXIT: {
+                        boolean success = saveFile(collectionPath, list);
+                        if (success)
+                            System.out.println("файл сохранён успешно");
+                        else
+                            System.out.println("не удалось сохранить файл");
+                        System.out.println("завершение программы");
+                        isExit = true;
+                        break;
+                    }
+                    case CMD_INFO: {
+                        System.out.println("тип " + list.getClass().getSimpleName());
+                        System.out.println("количество элементов " + list.size());
+                        System.out.println("дата инициализации " + initDate.toString());
+                        break;
+                    }
+                    case CMD_SHOW: {
+                        for (int i = 0; i < list.size(); i++) {
+                            House house = list.get(i);
+                            System.out.println(i + ": " + house.toString());
+                        }
+                        break;
+                    }
+                    case CMD_REMOVE: {
+                        if (!scanner.hasNext()) {
+                            System.out.println("ошибка: неправильный формат команды");
+                            continue;
+                        }
+                        try {
+                            int index = scanner.nextInt();
+                            if (index >= 0 && index < list.size()) {
+                                list.remove(index);
+                                list.sort(objAreaComparator);
+                                System.out.println("объект удалён");
+                            } else {
+                                System.out.println("ошибка: недопустимый индекс массива");
+                            }
+                        } catch (Exception e1) {
+                            try {
+                                String param = scanner.next();
+                                House house = new Gson().fromJson(param, House.class);
+                                boolean removed = list.remove(house);
+                                if (removed) {
+                                    list.sort(objAreaComparator);
+                                    System.out.println("объект удалён");
+                                } else
+                                    System.out.println("такой объект не найден");
+                            } catch (Exception e2) {
+                                System.out.println("ошибка: не получилось распознать объект из данных");
+                            }
+                        }
+                        scanner.nextLine();
+                        break;
+                    }
+
+                    case CMD_ADD: {
+                        try {
+                            String input = readMultiline();
+                            House house = new Gson().fromJson(input, House.class);
+                            if (house != null) {
+                                if (validateHouse(house)) {
+                                    list.add(house);
+                                    list.sort(objAreaComparator);
+                                    System.out.println("объект добавлен");
+                                } else {
+                                    System.err.println("инвалидно введено");
+                                }
+                            } else {
+                                System.err.println("ошибка: не получилось распознать объект из данных");
+                            }
+                        } catch (Exception e2) {
+                            System.err.println("ошибка при вводе данных");
+                        }
+                        break;
+                    }
+
+                    case CMD_ADD_IF_MIN: {
+                        try {
+                            String input = readMultiline();
+                            House house = new Gson().fromJson(input, House.class);
+                            if (house != null) {
+                                if (validateHouse(house) && house.area < list.get(0).area) {
+                                    list.add(house);
+                                    list.sort(objAreaComparator);
+                                    System.out.println("объект добавлен");
+                                } else {
+                                    System.err.println("инвалидно введено или площадь больше максимальной");
+                                }
+                            } else {
+                                System.err.println("ошибка: не получилось распознать объект из данных");
+                            }
+                        } catch (Exception e2) {
+                            System.err.println("ошибка при вводе данных");
+                        }
+                        break;
+                    }
+
+                    case CMD_INSERT: {
+                        try {
+                            int index = scanner.nextInt();
+                            String input = readMultiline();
+                            House house = new Gson().fromJson(input, House.class);
+                            list.add(index, house);
+                            list.sort(objAreaComparator);
+                            System.out.println("объект добавлен");
+                        } catch (Exception e2) {
+                            System.out.println("ошибка: не получилось распознать объект из данных");
+                        }
+                        break;
+                    }
+
+                    case CMD_SAVE: {
+                        boolean success = saveFile(collectionPath, list);
+                        if (success)
+                            System.out.println("файл сохранён успешно");
+                        else
+                            System.out.println("не удалось сохранить файл");
+                        break;
+                    }
+
+                    case CMD_HELP: {
+                        System.out.println("Информация о доступных командах:" +
+                                "\ninfo: вывести в стандартный поток вывода информацию о коллекции (тип, дата инициализации, количество элементов и т.д.)" +
+                                "\nshow: вывести в стандартный поток вывода все элементы коллекции в строковом представлении" +
+                                "\nremove {int index}: удалить элемент, находящийся в заданной позиции коллекции" +
+                                "\nadd {element}: добавить новый элемент в коллекцию" +
+                                "\ninsert {int index} {element}: добавить новый элемент в заданную позицию" +
+                                "\nremove {element}: удалить элемент из коллекции по его значению" +
+                                "\nsave: сохранить коллекцию в файл");
+                        break;
+                    }
+
+                    default: {
+                        System.out.println("неизвестная команда, попробуйте еще");
+                    }
+                }
+            }
+        } catch (NoSuchElementException e) {
+            System.out.println("Файл сохранен!");
+            saveFile(collectionPath, list);
+        }
 
         Scenario scenario = new Scenario() {
             @Override
@@ -151,5 +297,26 @@ public class Main {
         };
         //  scenario.Run();
     }
+
+    private static boolean validateHouse(House house) {
+        return house != null && house.area != 0 && !house.getName().isEmpty();
+    }
+
+    private static String readMultiline() {
+        StringBuilder builder = new StringBuilder();
+        String line = scanner.next();
+        builder.append(line);
+        line = scanner.nextLine();
+        while (scanner.hasNextLine()) {
+            line = scanner.nextLine();
+            if (line.isEmpty())
+                break;
+            else
+                builder.append(line);
+        }
+        return builder.toString();
+    }
+
+
 }
